Added resizing, (not working for walls entirely). Also fixed the tools to go into right place.

+++ src/WallPoint.cpp
@@ -24,7 +24,8 @@ along with Vectron.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "WallPoint.h"
 
-WallPoint::WallPoint( int newX, int newY ) {
+WallPoint::WallPoint( double newX, double newY ) {
     x = newX;
     y = newY;
+
 }
